Refactor project structure: remove unused files, update Makefile for cleanup commands, and change comment convention

+++ main.c
@@ -1,7 +1,7 @@
-#include "define.h"
+#include "global.h"
 
 int main(int argc, char *argv[]){
-    int input[BASE2_INSTRUCTION_LENGTH] = {0,0,0,1,1,0,1,1,2,9};
+    short int input[BASE2_INSTRUCTION_LENGTH] = {0,0,0,1,1,0,1,1,2,9};
     char *base4_output;
     base4_output = base4(input);
